requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0